tools/scripts/data_processing/Original_ARCIS_System/requirements_gps.txt
# ARCIS GPS Requirements for L76K Module Integration
# Install with: pip install -r requirements_gps.txt

# Core GPS dependencies
pyserial>=3.5
pynmea2>=1.19.0

# Military Grid Reference System (MGRS) support
mgrs>=1.4.0

# Existing ARCIS dependencies
ultralytics>=8.0.0
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0
matplotlib>=3.6.0
numpy>=1.21.0
Pillow>=9.0.0
PyYAML>=6.0
pygame>=2.1.0

# Optional: For enhanced GPS calculations
geopy>=2.3.0
pyproj>=3.4.0

# Development and debugging
ipython>=8.0.0 